+++ x/collection/keeper/supply_test.go
@@ -199,11 +199,17 @@ func (s *KeeperTestSuite) TestModifyContract() {
 			ctx, _ := s.ctx.CacheContext()
 
 			err := s.keeper.ModifyContract(ctx, contractID, s.vendor, changes)
-			if contractID == s.contractID {
-				s.Require().NoError(err)
-			} else {
+			if contractID != s.contractID {
 				s.Require().Error(err)
+				return
 			}
+			s.Require().NoError(err)
+
+			updated, err := s.keeper.GetContract(ctx, contractID)
+			s.Require().NoError(err)
+			s.Require().Equal(changes[0].Value, updated.Name)
+			s.Require().Equal(changes[1].Value, updated.BaseImgUri)
+			s.Require().Equal(changes[2].Value, updated.Meta)
 		})
 	}
 }
@@ -229,11 +235,18 @@ func (s *KeeperTestSuite) TestModifyTokenClass() {
 				ctx, _ := s.ctx.CacheContext()
 
 				err := s.keeper.ModifyTokenClass(ctx, contractID, classID, s.vendor, changes)
-				if contractID == s.contractID && classID == s.nftClassID {
-					s.Require().NoError(err)
-				} else {
+				if contractID != s.contractID || classID != s.nftClassID {
 					s.Require().Error(err)
+					return
 				}
+				s.Require().NoError(err)
+
+				updated, err := s.keeper.GetTokenClass(ctx, contractID, classID)
+				s.Require().NoError(err)
+				casted, ok := updated.(*collection.NFTClass)
+				s.Require().True(ok, "must be nft class")
+				s.Require().Equal(changes[0].Value, casted.Name)
+				s.Require().Equal(changes[1].Value, casted.Meta)
 			})
 		}
 	}
@@ -261,11 +274,16 @@ func (s *KeeperTestSuite) TestModifyNFT() {
 				ctx, _ := s.ctx.CacheContext()
 
 				err := s.keeper.ModifyNFT(ctx, contractID, tokenID, s.vendor, changes)
-				if contractID == s.contractID && tokenID == validTokenID {
-					s.Require().NoError(err)
-				} else {
+				if contractID != s.contractID || tokenID != validTokenID {
 					s.Require().Error(err)
+					return
 				}
+				s.Require().NoError(err)
+
+				updated, err := s.keeper.GetNFT(ctx, contractID, tokenID)
+				s.Require().NoError(err)
+				s.Require().Equal(changes[0].Value, updated.Name)
+				s.Require().Equal(changes[1].Value, updated.Meta)
 			})
 		}
 	}